Make visbileBinding work for InfoBoxView
REDMINE-20438
