+++ app/assets/stylesheets/pageflow/ui/forms.scss
@@ -219,7 +219,7 @@ textarea.short {
   text-decoration: underline;
 }
 
-.input-hidden_via_binding {
+.hidden_via_binding {
   display: none;
 }
