+++ package/src/ui/views/mixins/inputView.js
@@ -1,6 +1,7 @@
 import _ from 'underscore';
 
 import {attributeTranslationKeys, findTranslation, translationKeysWithSuffix} from '../../utils/i18nUtils';
+import {attributeBinding} from './attributeBinding';
 
 /**
  * Mixin for input views handling common concerns like labels,
@@ -131,6 +132,8 @@ import {attributeTranslationKeys, findTranslation, translationKeysWithSuffix} fr
  * @mixin
  */
 export const inputView = {
+  ...attributeBinding,
+
   ui: {
     label: 'label',
     labelText: 'label .name',
@@ -262,54 +265,7 @@ export const inputView = {
   },
 
   updateVisible: function() {
-    this.$el.toggleClass('input-hidden_via_binding',
+    this.$el.toggleClass('hidden_via_binding',
                          this.getBooleanAttributBoundOption('visible') === false);
-  },
-
-  setupBooleanAttributeBinding(optionName, updateMethod) {
-    this.setupAttributeBinding(optionName, updateMethod, Boolean);
-  },
-
-  getBooleanAttributBoundOption(optionName) {
-    return this.getAttributeBoundOption(optionName, Boolean);
-  },
-
-  setupAttributeBinding: function(optionName, updateMethod, normalize = value => value) {
-    const binding = this.options[`${optionName}Binding`];
-    const view = this;
-
-    if (binding) {
-      _.flatten([binding]).forEach(attribute => {
-        this.listenTo(this.model, 'change:' + attribute, update);
-      });
-    }
-
-    update();
-
-    function update() {
-      updateMethod.call(view, view.getAttributeBoundOption(optionName, normalize));
-    }
-  },
-
-  getAttributeBoundOption(optionName, normalize = value => value) {
-    const binding = this.options[`${optionName}Binding`];
-    const bindingValueOptionName = `${optionName}BindingValue`;
-
-    const value = Array.isArray(binding) ?
-                  binding.map(attribute => this.model.get(attribute)) :
-                  this.model.get(binding);
-
-    if (bindingValueOptionName in this.options) {
-      return value === this.options[bindingValueOptionName];
-    }
-    else if (typeof this.options[optionName] === 'function') {
-      return normalize(this.options[optionName](value));
-    }
-    else if (optionName in this.options) {
-      return normalize(this.options[optionName]);
-    }
-    else if (binding) {
-      return normalize(value);
-    }
   }
 };
